paypal and other changes

+++ src/main/java/hr/algebra/java/web/lego/legostorewithdata/repository/CartItemRepository.java
@@ -7,7 +7,7 @@ import java.util.List;
 public interface CartItemRepository {
 
     List<CartItem> getCartItems(String username);
-    CartItem getCartItem(Long id);
+    CartItem getCartItem(int id);
     void updateCartItem(CartItem cartItem);
 
     void deleteCartItem(CartItem cartItem);
